sql upload, bugfixes on py
