+++ pymenu/main.py
@@ -1,12 +1,13 @@
-from Users import *
+from users import users_menu
 
 
 def main_menu():
+    choice = 0
     print("1. users")
     print("2. employees")
     print("3. cabins")
     print("4. shop")
-    print("5. exit")
+    print("0. EXIT")
     choice = input("Enter your choice: ")
     if choice == "1":
         users_menu()
@@ -15,9 +16,11 @@ def main_menu():
     elif choice == "3":
         cabins_menu()
     elif choice == "4":
-        employees_menu()
-    elif choice == "5":
         shop_menu()
+    elif choice == "0":
+        print("Bye!")
+        print("designed by uni2")
+        exit()
     else:
         print("Invalid choice")
         main_menu()
