+++ pymenu/users.txt
@@ -1,2 +1,4 @@
-1,2,3,4,5
-erwq,eqw,qwe,ewq,qwq
+erlantz garate ei@uwu.com uwu 112
+3421 423 432 432 911
+uwu uwu uwu uwu uwu
+0 0 0 0 0
